+++ src/PopDesign.Application/Dtos/TaxasMarketplaceDto.cs
@@ -6,7 +6,7 @@ public class TaxasMarketplaceDto
     public Guid IdMarketplace { get; set; }
     public decimal ValorInicial { get; set; }
     public decimal ValorFinal { get; set; }
-    public int Comissao { get; set; }
+    public decimal Comissao { get; set; }
     public decimal TaxaFixa { get; set; }
     public string MarketplaceNome { get; set; } = string.Empty;
 }
