+++ src/PopDesign.Domain/Entities/TaxasMarketplace.cs
@@ -6,7 +6,7 @@ public class TaxasMarketplace
     public Guid IdMarketplace { get; set; }
     public decimal ValorInicial { get; set; }
     public decimal ValorFinal { get; set; }
-    public int Comissao { get; set; }
+    public decimal Comissao { get; set; }
     public decimal TaxaFixa { get; set; }
 
     public virtual Marketplace Marketplace { get; set; } = null!;
